+++ src/froGH/froGH/Utils/Utilities.cs
@@ -43,66 +43,58 @@ namespace froGH.Utils
                 vertices[j] = edgeTips.I == tvIndex ? edgeTips.J : edgeTips.I;
             }
 
-
-            //
-            //
-            // this part should be rewritten to consider naked edges cases (right now it's not working as it should)
-            //
-
             faces = null;
 
-
             // look for faces only if there is more than 1 edge
-            if (edges.Length > 1)
+            if (edges.Length == 1) return;
+
+            // find neighbour faces in edges order
+            // try this: pick first face index as face shared from first and second edge
+            // for the following edges pick the index (of the two) that is different from the previous iteration
+            faces = new int[edges.Length];
+            List<int> facesList = new List<int>();
+
+
+            // first edge/face index
+            int[] connectedFaces;
+            int faceCount = 0;
+            // find connected faces indexes to first edge
+            connectedFaces = M.TopologyEdges.GetConnectedFaces(edges[0]);
+            if (connectedFaces.Length != 0)
             {
-                // find neighbour faces in edges order
-                // try this: pick first face index as face shared from first and second edge
-                // for the following edges pick the index (of the two) that is different from the previous iteration
-                faces = new int[edges.Length];
-                List<int> facesList = new List<int>();
-
-
-                // first edge/face index
-                int[] cFaces;
-                int faceCount = 0;
-                // find connected faces indexes to first edge
-                cFaces = M.TopologyEdges.GetConnectedFaces(edges[0]);
-                if (cFaces.Length != 0)
+                for (int j = 0; j < connectedFaces.Length; j++)
                 {
-                    for (int j = 0; j < cFaces.Length; j++)
+                    // find edges for j-th connected face
+                    int[] fEdges = M.TopologyEdges.GetEdgesForFace(connectedFaces[j]);
+                    // if face contains also index of second edge then we found our first face
+                    if (fEdges.Contains(edges[1]))
                     {
-                        // find edges for j-th connected face
-                        int[] fEdges = M.TopologyEdges.GetEdgesForFace(cFaces[j]);
-                        // if face contains also index of second edge then we found our first face
-                        if (fEdges.Contains(edges[1]))
-                        {
-                            facesList.Add(cFaces[j]);
-                            faceCount = 1;
-                            //faces[0] = cFaces[j];
-                            break;
-                        }
+                        facesList.Add(connectedFaces[j]);
+                        faceCount = 1;
+                        break;
                     }
                 }
-                // all other edges
-                for (int j = 1; j < edges.Length; j++)
-                {
-                    cFaces = M.TopologyEdges.GetConnectedFaces(edges[j]);
-                    // if it's a clothed edge
-                    if (cFaces.Length > 1)
-                    {
-                        facesList.Add(cFaces[0] != facesList[faceCount - 1] ? cFaces[0] : cFaces[1]);
-                        faceCount++;
-                    }
-                    // else, if naked...
-                    else if (cFaces.Length > 0 && cFaces[0] != facesList[faceCount - 1])
-                    {
-                        facesList.Add(cFaces[0]);
-                        faceCount++;
-                    }
-                }
-
-                faces = facesList.ToArray();
             }
+            // all other edges
+            for (int j = 1; j < edges.Length; j++)
+            {
+                connectedFaces = M.TopologyEdges.GetConnectedFaces(edges[j]);
+                // if it's a clothed edge
+                if (connectedFaces.Length > 1)
+                {
+                    facesList.Add(connectedFaces[0] != facesList[faceCount - 1] ? connectedFaces[0] : connectedFaces[1]);
+                    faceCount++;
+                }
+                // else, if naked...
+                else if (connectedFaces.Length > 0 && connectedFaces[0] != facesList[faceCount - 1])
+                {
+                    facesList.Add(connectedFaces[0]);
+                    faceCount++;
+                }
+            }
+
+            faces = facesList.ToArray();
+
         }
 
         public static Mesh MakeSphereMesh(Point3d P, double R, int step)